Add files via upload
The two python files (Njc_Labs_Internship_program_part1.py and Njc_Labs_Internship_program_part2.py ) are actually just one file, but for simplicity, I have segregated the code into two python files. in this file, The code will first check the connection (verify username and password) and connect to the server (provided it does not give any Error). Then it creates the Database and sees all the database names on the server(this is optional).