Add files via upload
in this file, Code checks the connection (if provided username, password, and database name are correct). Then it creates the table named "movie" and inserts the data in this table. As said in the instructions, it queries the movie details ( name, actor, actress, director, year of release) and gives all the movie names if we pass the lead_actor as a parameter).